refactoring / update README.md

+++ .files/scripts/upgrade.sh
@@ -9,10 +9,14 @@ branch="$(git rev-parse --abbrev-ref HEAD)"
 [[ $branch == "master" ]] || git checkout master
 
 # pull changes in master branch
-git pull
+GIT_RESULT=$(git pull)
 
-# print new line
-echo ""
+# print result
+echo "${GIT_RESULT}"
+
+if [[ ${GIT_RESULT} == "Already up to date." ]] ; then
+    exit 0
+fi
 
 # store date of update
 echo "$(date +%s)" > "${DOTFILES_LAST_PULL_PATH}"
